persona_chatbot-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
alembic
mako>=1.3.0
email-validator
python-dotenv
groq>=0.22.0
typing-extensions>=4.5.0
certifi>=2024.2.2
authlib
httpx